Fixing geometry/adapted/cuda/float return types. Adding cast function to geometry.

+++ particle/geometry/adapted/cuda/float2.hpp
@@ -45,19 +45,19 @@ namespace geometry
       struct apply<Float2, 0>
       {
         PARTICLE_STATIC_FUNCTION
-        int& call(float2& f)
+        float& call(float2& f)
         {
           return f.x;
         }
 
         PARTICLE_STATIC_FUNCTION
-        const int& call(const float2& f)
+        const float& call(const float2& f)
         {
           return f.x;
         }
 
         PARTICLE_STATIC_FUNCTION
-        int call(float2&& f)
+        float call(float2&& f)
         {
           return f.x;
         }
@@ -67,19 +67,19 @@ namespace geometry
       struct apply<Float2, 1>
       {
         PARTICLE_STATIC_FUNCTION
-        int& call(float2& f)
+        float& call(float2& f)
         {
           return f.y;
         }
 
         PARTICLE_STATIC_FUNCTION
-        const int& call(const float2& f)
+        const float& call(const float2& f)
         {
           return f.y;
         }
 
         PARTICLE_STATIC_FUNCTION
-        int call(float2&& f)
+        float call(float2&& f)
         {
           return f.y;
         }
