+++ particle/geometry/adapted/cuda/float3.hpp
@@ -45,19 +45,19 @@ namespace geometry
       struct apply<Float3, 0>
       {
         PARTICLE_STATIC_FUNCTION
-        int& call(float3& f)
+        float& call(float3& f)
         {
           return f.x;
         }
 
         PARTICLE_STATIC_FUNCTION
-        const int& call(const float3& f)
+        const float& call(const float3& f)
         {
           return f.x;
         }
 
         PARTICLE_STATIC_FUNCTION
-        int call(float3&& f)
+        float call(float3&& f)
         {
           return f.x;
         }
@@ -67,19 +67,19 @@ namespace geometry
       struct apply<Float3, 1>
       {
         PARTICLE_STATIC_FUNCTION
-        int& call(float3& f)
+        float& call(float3& f)
         {
           return f.y;
         }
 
         PARTICLE_STATIC_FUNCTION
-        const int& call(const float3& f)
+        const float& call(const float3& f)
         {
           return f.y;
         }
 
         PARTICLE_STATIC_FUNCTION
-        int call(float3&& f)
+        float call(float3&& f)
         {
           return f.y;
         }
@@ -89,19 +89,19 @@ namespace geometry
       struct apply<Float3, 2>
       {
         PARTICLE_STATIC_FUNCTION
-        int& call(float3& f)
+        float& call(float3& f)
         {
           return f.z;
         }
 
         PARTICLE_STATIC_FUNCTION
-        const int& call(const float3& f)
+        const float& call(const float3& f)
         {
           return f.z;
         }
 
         PARTICLE_STATIC_FUNCTION
-        int call(float3&& f)
+        float call(float3&& f)
         {
           return f.z;
         }
